==> content/quantecol/chapters/04-env_dist.Rmd <==
---
date: "2021-01-01"
draft: false
excerpt: null
links:
  - icon: images
    icon_pack: fas
    name: Environmental Distance Slides
    url: /pdf/BCB743/BCB743_04_environmental_distance.pdf
  - icon: file-csv
    icon_pack: fa
    name: Euclidian Distance Demo Data
    url: /BCB743/seaweed/Euclidian_distance_demo_data_xyz.csv
  - icon: file-csv
    icon_pack: fa
    name: Euclidian Distance Demo Env Data
    url: /BCB743/seaweed/Euclidian_distance_demo_data_env.csv
  - icon: file-csv
    icon_pack: fa
    name: Sites Data
    url: /BCB743/seaweed/sites.csv
subtitle: ""
title: 4. Environmental Distance
weight: 6
---

<!--- # Topic 4: Environmental distance --->

## Set up the analysis environment

```{r message=FALSE, warning=FALSE}
library(vegan)
library(ggplot2)
library(geodist) # for calculating geographical distances between lats/lons
library(ggpubr) # to arrange the multipanel graphs

# setting up a 'root' file path so I don't have to keep doing it later...
root <- "../../../static/BCB743/"
```

## Toy data

The toy data have arbitrary columns to demonstrate the Euclidean distance calculation in the lecture slides:

$$ d(a,b) = \sqrt{(a_x - b_x)^2 + (a_y - b_y)^2 + (a_z - b_z)^2} $$

The distance is found between every pair of sites named ***a*** to ***g*** whose locations are marked by the 'coordinates' $x$, $y$, and $z$---i.e. this is an example of 3-dimensional data (a space or volume, as opposed to 2D data situated on a $x$, $y$ place). We might also call each coordinate a 'variable' (sometimes called a 'dimension') and hence we have multivariate or multidimensional data.

Let's load the dataset and find the size of the dataframe:

```{r}
xyz <- read.csv(paste0(root, "seaweed/Euclidian_distance_demo_data_xyz.csv"))
dim(xyz)
```

There are seven rows and four columns.

The data look like:

```{r}
xyz
```

The first column contains the site names and it must be excluded from subsequent calculations. The remaining three columns will be used below.

Calculate the Euclidean distance using **vegan**'s `vegdist()` function and view the lower triangle with the diagonal:

```{r}
xyz_euc <- round(vegdist(xyz[, 2:4], method = "euclidian",
                         upper = FALSE, diag = TRUE), 4)
# selected only cols 2, 3 and 4
xyz_euc
```

Convert to a dataframe and view it:

```{r}
xyz_df <- as.data.frame(round(as.matrix(xyz_euc), 4))
xyz_df
```

Distance matrices have the same properties as dissimilarity matrices, i.e.:

i.  The distance matrix is square (number rows = number columns).

ii. The diagonal is filled with 0.

iii. The matrix is symmetrical---it is comprised of symmetrical upper and lower triangles.

In terms of the meaning of the cell values, their interpretation is also analogous with that of the species dissimilarities. A value of 0 means the properties of the sites (or sections, plots, transects, quadrats, etc.) in terms of their environmental conditions are identical (this is always the case the the diagonal). The larger the number (which may be \>1) the more different sites are in terms of their environmental conditions.

Since each column, $x$, $y$, and $z$, is a variable, we can substitute them for *actual* variables or properties of the environment within which species are present. Let's load such data (again fictitious):

```{r}
env_fict <- read.csv(paste0(root, "seaweed/Euclidian_distance_demo_data_env.csv"))
head(env_fict, 2) # print first two rows only
```

These are the same data as in `Euclidian_distance_demo_data_xyz.csv` but I simply renamed the columns to names of the variables temperature, depth, and light intensity. I won't repeat the analysis here as the output remains the same.

Now apply `vegdist()` as before. The resultant distances are called 'environmental distances'.

Let us now use some real data.

## A look at the seaweed environmental data

These data accompany the analysis of the South African seaweed flora (see Smit et al. 2017).

```{r}
load(paste0(root, "seaweed/SeaweedEnv.RData"))

# lets look at the data
dim(env)
```

We see that the data have 58 rows and 18 columns... the same number of rows as the `seaweed.csv` data. What is in the first five rows?

```{r}
round(env[1:5, 1:5], 4)
```

And the last five rows?

```{r}
round(env[(nrow(env) - 5):nrow(env), (ncol(env) - 5):ncol(env)], 4)
```

So, each of the rows corresponds to a site (i.e. each of the coastal sections), and the columns each contains an environmental variable. The names of the environmental variables are:

```{r}
colnames(env)
```

As we have seen, there are 18 variables (or dimensions). These data are truly multidimensional in a way that far exceeds our brains' limited ability to spatially visualise. For mathematicians these data define an 18-dimensional space, but all we can do is visualise 3-dimensions.

We select only some of the thermal variables; the rest are collinear with some of the ones I import:

```{r}
  env1 <- dplyr::select(env, febMean, febRange, febSD, augMean,
                    augRange, augSD, annMean, annRange, annSD)
```

## Calculate *z*-scores

Here we need to do something new that was not necessary with the toy data. We calculate *z*-scores, and the process is called 'standardisation'. Standardisation is necessary when the variables are measured in diffrent units---e.g. the unit for temperature is °C whereas Ch-*a* is measured in mg Chl-*a*/m3.

```{r}
E1 <- round(decostand(env1, method = "standardize"), 4)
E1[1:5, 1:5]
```

## Calculate Euclidian distance

```{r}
E1_euc <- round(vegdist(E1, method = "euclidian", upper = TRUE), 4)
E1_df <- as.data.frame(as.matrix(E1_euc))
E1_df[1:10, 1:10]
```

We already know how to read this matrix. Let's plot it as a function of the coastal section's number:

```{r}
ggplot(data = E1_df, (aes(x = 1:58, y = `1`))) +
  geom_line() +
  xlab("Coastal section, west to east") +
  ylab("Environmental distance")
```

We will explore distance and dissimilarity matrices in more detail in later sections.

## Euclidian distances of geographical data

When we calculate Euclidian distances between geographic lat/lon coordinate, the relationship between sections will be the same (but scaled) as actual geographic distances.

```{r}
geo <- read.csv(paste0(root, "seaweed/sites.csv"))
dim(geo)
```

```{r}
head(geo)
```

<!--- The first column contains nonsense and must be removed: --->

Calculate geographic distances (in meters) between coordinate pairs:

```{r}
dists <- geodist(geo, paired = TRUE, measure = "geodesic")
dists_df <- as.data.frame(as.matrix(dists))
colnames(dists_df) <- seq(1:58)
dists_df[1:5, 1:5]
```

```{r}
plt1 <- ggplot(data = dists_df, (aes(x = 1:58, y = `1`/1000))) +
  geom_line() +
  xlab("Coastal section, west to east") +
  ylab("Distance (km)") +
  ggtitle("Actual geographic distance")
```

```{r}
dists_euc <- vegdist(geo, method = "euclidian")
dists_euc_df <- round(as.data.frame(as.matrix(dists_euc)), 4)
dists_euc_df[1:5, 1:5]
```

```{r}
plt2 <- ggplot(data = dists_euc_df, (aes(x = 1:58, y = `1`))) +
  geom_line() +
  xlab("Coastal section, west to east") +
  ylab("Euclidian distance") +
  ggtitle("Euclidian distance")
```

```{r}
ggarrange(plt1, plt2, nrow = 2)
```

## References

Smit, A. J., Bolton, J. J., & Anderson, R. J. (2017). Seaweeds in two oceans: beta-diversity. Frontiers in Marine Science, 4, 404.
