==> content/quantecol/chapters/06-correlations.Rmd <==
---
date: "2021-01-01"
draft: false
excerpt: null
links:
  - icon: images
    icon_pack: fas
    name: Correlations Slides
    url: /pdf/BCB743/BCB743_06_correlations.pdf
  - icon: file-export
    icon_pack: fa
    name: Doubs.RData
    url: /BCB743/NEwR-2ed_code_data/NEwR2-Data/Doubs.RData
subtitle: ""
title: "6. Correlations and Associations"
weight: 8
editor_options: 
  chunk_output_type: console
output: 
  html_document: 
    toc: yes
    fig_caption: yes
---

<!--- # Topic 6: Correlations and associations --->

## Set-up the analysis environment

```{r message=FALSE, warning=FALSE}
library(tidyverse)
library(vegan)
library(Hmisc) # for rcorr()

# setting up a 'root' file path so I don't have to keep doing it later...
root <- "../../../static/BCB743/"
```

## The Doubs River data

The background to the data is described by David Zelený on his excellent [website](https://www.davidzeleny.net/anadat-r/doku.php/en:data:doubs) and in the book **Numerical Ecology with R** by Borcard et al. (2011). These data are a beautiful example of how gradients structure biodiversity. It will be in your own interest to fully understand how the various environmental factors used as explanatory variables vary along a riverine gradient from the source to the terminus of the river.

### Correlations between environmental variables

```{r}
load(paste0(root, "NEwR-2ed_code_data/NEwR2-Data/Doubs.RData"))

head(env, 5)
```

We use [correlations](https://ajsmit.netlify.app/workshops/quantecol/chapters/06-correlations/) to establish how the environmental variables relate to one another across the sample sites. We do not need to standardise as one would do for the calculation of Euclidian distances, but in some instances data transformations might be necessary:

```{r}
env_cor <- round(cor(env), 2)
env_cor
```

Or if we want to see the associated *p*-values to establish a statistical significance:

```{r}
rcorr(as.matrix(env))
```

We can also do a visual exploration (see Question 1, below).

<!-- ```{r} -->
<!-- library(ggcorrplot) -->
<!-- ggcorrplot(env_cor, type = 'upper', outline.col = "white", -->
<!--            colors = c("#00AFBB", "white", "#FC4E07"), -->
<!--            lab = TRUE) -->
<!-- ``` -->

### Assignment 3 Questions

> **Question 1:** Create a plot of pairwise correlations.
>
> **Question 2:** Name to two top positive and two top negative *statistically-significant* correlations.
>
> **Question 3:** For each, discuss the mechanism behind the relationships. Why do these relationships exist?

### Association between species

The Doubs River fish species dataset is an example of abundance data and it will serve well to examine the properties of an association matrix:

```{r}
head(spe)
```

In order to calculate an association matrix for the fish species we first need to **transpose** the data:

```{r}
spe_t <- t(spe)
```

### Assignment 3 Questions

> **Question 4:** Why do we need to transpose the data?
>
> **Question 5:** What are the properties of a transposed species table?

Now we can calculate the association matrix:

```{r}
spe_assoc1 <- vegdist(spe_t, method = "jaccard")
 # display only a portion of the data...
as.matrix((spe_assoc1))[1:10, 1:10]
```

```{r}
spe_assoc2 <- vegdist(spe_t, method = "jaccard", binary = TRUE)
as.matrix((spe_assoc2))[1:10, 1:10]
```

### Assignment 3 Questions

> **Question 6:** What are the properties of an association matrix? How do these properties differ from that of a i) species dissmilarity matrix and from a ii) correlation matrix?
>
> **Question 7:** What is the difference between `spp_assoc1` and `spp_assoc2`? Is the information contained in each markedly different from the other?
>
> **Question 8:** Explain the kind of insight we are able to glean from a species association matrix.

Submit a R script wherein you provide answers to Questions 1--8 by no later than 8:00 tomorrow. Label the script as follows: **`BCB743_<Name>_<Surname>_Assignment_3.R`**, e.g. **`BCB743_AJ_Smit_Assignment_3.R`**.

## References

Borcard, D., Gillet, F. & Legendre, P. (2011) Numerical Ecology with R. Springer.
