==> content/quantecol/chapters/07-ordination.Rmd <==
---
date: "2021-01-01"
draft: false
excerpt: null
links:
  - icon: images
    icon_pack: fas
    name: Ordination Slides
    url: /pdf/BCB743/BCB743_07_ordination.pdf
  - icon: file-pdf
    icon_pack: far
    name: Vegan -- an introduction to ordination
    url: /pdf/BCB743/Oksanen_intro-vegan.pdf
subtitle: ""
title: "7. Ordination"
weight: 9
---

<!--- # Topic 7: Ordination --->

Refer to the lecture slides.

Typically, ordination is classified into **indirect gradient analysis** and **direct gradient analysis**. These will be covered in the next few chapters.

## Indirect gradient analysis

We apply indirect gradient analysis when the gradients are unknown *a priori* and we do not have environmental data related to the species. Gradients or other influences that structure species in space are therefore inferred from the species composition data only. The communities thus tell us that gradients are present (or not), but it might not offer insight into the identity of the structuring gradients.

Please refer to:

* [Principal Component Analysis](/quantecol/chapters/08-pca/)
* [Correspondence Analysis](/quantecol/chapters/09-ca/)
* [Principal Coordinate Analysis](/quantecol/chapters/10-pcoa/)
* [non-Metric Multidimensional Scaling](/quantecol/chapters/11-nmds/)

## Direct gradient analysis

Any ordination approach in which important gradients are hypothesised. Likely evidence for the existence of gradients is measured and captured in a complementary environmental dataset that has the same spatial structure (rows) as the species dataset. Direct gradient analysis is performed using linear or non-linear regression approaches that relate the ordinations performed on the species and their matching environmental variables.

* Redundancy Analysis
* Constrained Correspondence Analysis
* [distance-based Redundancy Analysis](/quantecol/chapters/12-constrained_ordination/)
